+++ inst/app/radiocarbon/global.R
@@ -1,0 +1,8 @@
+## +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+## Title:   Radiocarbon Calibration
+## Authors: Nicolas Frerebeau, Université Bordeaux Montaigne (France)
+## Contact: nicolas.frerebeau@u-bordeaux-montainge.fr
+##++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+## Set shiny settings ==========================================================
+options(shiny.maxRequestSize = 30*1024^2)